Merge branch 'develop' into feature-25478

+++ Classes/Domain/Model/Album.php
@@ -2,7 +2,7 @@
 /***************************************************************
 *  Copyright notice
 *
-*  (c) 2010 Michael Knoll <mimi@kaktusteam.de>
+*  (c) 2010-2011 Michael Knoll <mimi@kaktusteam.de>
 *  			Daniel Lienert <daniel@lienert.cc>
 *  			
 *  All rights reserved
@@ -326,22 +326,21 @@ class Tx_Yag_Domain_Model_Album extends Tx_Extbase_DomainObject_AbstractEntity {
     public function setThumb(Tx_Yag_Domain_Model_Item $thumb) {
         $this->thumb = $thumb;
     }
-    
-    
 
-    /**
-     * Getter for thumb
-     *
-     * @return Tx_Yag_Domain_Model_Item Holds thumbnail for this album
-     */
-    public function getThumb() {
-    	// we need this because we check the class name in itemViewHelper
-        if(get_class($this->thumb) === 'Tx_Extbase_Persistence_LazyLoadingProxy') {
-        	return $this->thumb->_loadRealInstance();	
-        } else {
-        	return $this->thumb;
-        }
-    }
+
+	/**
+	 * Getter for thumb
+	 *
+	 * @return Tx_Yag_Domain_Model_Item Holds thumbnail for this album
+	 */
+	public function getThumb() {
+		// we need this because we check the class name in itemViewHelper
+		if (get_class($this->thumb) === 'Tx_Extbase_Persistence_LazyLoadingProxy') {
+			return $this->thumb->_loadRealInstance();
+		} else {
+			return $this->thumb;
+		}
+	}
     
     
 
@@ -436,19 +435,41 @@ class Tx_Yag_Domain_Model_Album extends Tx_Extbase_DomainObject_AbstractEntity {
 	 */
 	public function delete($deleteItems = true) {
 		// To avoid complications, we first of all delete thumb
-		$this->thumb->delete();
-		$this->thumb = null;
+		$this->deleteThumb();
+		
 		if ($deleteItems) {
-			foreach ($this->items as $item) { /* @var $item Tx_Yag_Domain_Model_Item */
-				$item->delete();
-			}
+			$this->deleteAllItems();
 		}
+
 		$this->gallery->setThumbAlbumToTopOfAlbums();
 		$albumRepository = t3lib_div::makeInstance('Tx_Yag_Domain_Repository_AlbumRepository');
 		$albumRepository->remove($this);
 	}
-	
-	
+
+
+
+	/**
+	 * @return void
+	 */
+	public function deleteThumb() {
+		if($this->thumb && is_object($this->thumb)) {
+			$this->thumb->delete();
+			$this->thumb = null;
+		}
+	}
+
+
+
+	/**
+	 * @return void
+	 */
+	public function deleteAllItems() {
+		foreach ($this->items as $item) { /* @var $item Tx_Yag_Domain_Model_Item */
+				$item->delete();
+		}
+	}
+
+
 	
 	/**
 	 * Sets thumbnail to first item of items in this album
@@ -474,6 +495,63 @@ class Tx_Yag_Domain_Model_Album extends Tx_Extbase_DomainObject_AbstractEntity {
         } else {
         	return 0;
         }
-	}	
+	}
+
+
+
+    /**
+     * Updates sorting of items of this album.
+     *
+     * Sorting of items of this album is updated by given sorting field
+     * and sorting direction. Sorting field must be one field of item,
+     * sorting direction must be 1 = ASC or -1 = DESC.
+     *
+     * @param string $sortingField Field of item to be used for sorting
+     * @param string $sortingDirection Sorting direction to be used for sorting.
+     * @return void
+     */
+    public function updateSorting($sortingField, $sortingDirection) {
+        $itemRepository = t3lib_div::makeInstance('Tx_Yag_Domain_Repository_ItemRepository'); /* @var $itemRepository Tx_Yag_Domain_Repository_ItemRepository */
+        $sortedItems = $itemRepository->getSortedItemsByAlbumFieldAndDirection($this, $sortingField, $sortingDirection);
+        $this->items = new Tx_Extbase_Persistence_ObjectStorage();
+        foreach($sortedItems as $item) {
+            $this->items->attach($item);
+        }
+    }
+
+
+
+    /**
+     * Returns maximum sorting of items within this album
+     * 
+     * @return int
+     */
+    public function getMaxSorting() {
+        $itemRepository = t3lib_div::makeInstance('Tx_Yag_Domain_Repository_ItemRepository'); /* @var $itemRepository Tx_Yag_Domain_Repository_ItemRepository */
+        $maxSortingItem = $itemRepository->getItemWithMaxSortingForAlbum($this);
+        if (count($maxSortingItem) > 0) {
+            return $maxSortingItem[0]->getSorting();
+        } else {
+            return 0;
+        }
+    }
+
+
+
+    /**
+     * Checks, whether an image for a given filehash is included in album
+     *
+     * @param string $fileHash MD5 hash of item to be checked to be in this album
+     * @return bool True, if image is included
+     */
+    public function containsItemByHash($fileHash) {
+        foreach($this->items as $item) { /* @var $item Tx_Yag_Domain_Model_Item */
+            if ($item->getFilehash() == $fileHash) {
+                return true;
+            }
+        }
+        return false;
+    }
+
 }
 ?>
